chore: move icons to types file
