+++ packages/blade/src/components/Icons/_Svg/Circle/Circle.web.tsx
@@ -1,5 +1,5 @@
 import type { ReactElement } from 'react';
-import type { CircleProps } from './Circle.d';
+import type { CircleProps } from './types';
 
 const Circle = ({ cx, cy, x, y, r, fill, stroke }: CircleProps): ReactElement => {
   return <circle cx={cx} cy={cy} x={x} y={y} r={r} fill={fill} stroke={stroke} />;
