+++ packages/blade/src/components/Icons/_Svg/ClipPath/ClipPath.native.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 import { ClipPath as ClipPathNative } from 'react-native-svg';
-import type { ClipPathProps } from './ClipPath.d';
+import type { ClipPathProps } from './types';
 
 const ClipPath = ({ children, id }: ClipPathProps): ReactElement => {
   return <ClipPathNative id={id}>{children}</ClipPathNative>;
