+++ packages/blade/src/components/Icons/_Svg/ClipPath/ClipPath.web.tsx
@@ -1,5 +1,5 @@
 import type { ReactElement } from 'react';
-import type { ClipPathProps } from './ClipPath.d';
+import type { ClipPathProps } from './types';
 
 const ClipPath = ({ children, id }: ClipPathProps): ReactElement => {
   return <clipPath id={id}>{children}</clipPath>;
