+++ packages/blade/src/components/Icons/_Svg/ClipPath/types.ts
