+++ packages/blade/src/components/Icons/_Svg/Defs/Defs.native.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 import { Defs as DefsNative } from 'react-native-svg';
-import type { DefsProps } from './Defs.d';
+import type { DefsProps } from './types';
 
 const Defs = ({ children }: DefsProps): ReactElement => {
   return <DefsNative>{children}</DefsNative>;
