+++ packages/blade/src/components/Icons/_Svg/Defs/Defs.web.tsx
@@ -1,5 +1,5 @@
 import type { ReactElement } from 'react';
-import type { DefsProps } from './Defs.d';
+import type { DefsProps } from './types';
 
 const Defs = ({ children }: DefsProps): ReactElement => {
   return <defs>{children}</defs>;
