+++ packages/blade/src/components/Icons/_Svg/G/types.ts
