+++ packages/blade/src/components/Icons/_Svg/Path/types.ts
