+++ packages/blade/src/components/Icons/_Svg/Rect/Rect.web.tsx
@@ -1,5 +1,5 @@
 import type { ReactElement } from 'react';
-import type { RectProps } from './Rect.d';
+import type { RectProps } from './types';
 
 const Rect = ({ height, width, rx, ry, x, y, fill }: RectProps): ReactElement => {
   return <rect height={height} width={width} rx={rx} ry={ry} x={x} y={y} fill={fill} />;
