+++ packages/blade/src/components/Icons/_Svg/Rect/types.ts
